refactor: move useThrottle to src/hook/use-throttle.ts
