+++ src/components/page/Preference/components/model/PreferenceTabs/tabs/StatisticsTab/hook/index.ts
@@ -5,7 +5,8 @@ import {
   events,
 } from '@/lib/bindings'
 import { UnlistenFn } from '@tauri-apps/api/event'
-import { useEffect, useState, useRef } from 'react'
+import { useEffect, useState } from 'react'
+import { useThrottle } from '@/hooks/use-throttle'
 
 type ReturnProps = {
   assetRegistrationAreaChartData: AssetRegistrationStatistics[]
@@ -16,30 +17,6 @@ type ReturnProps = {
   avatars: number
   avatarWearables: number
   worldObjects: number
-}
-
-const useThrottle = <T>(value: T, delay: number): T => {
-  const [throttledValue, setThrottledValue] = useState<T>(value)
-  const lastUpdateTime = useRef<number>(Date.now())
-
-  useEffect(() => {
-    const now = Date.now()
-    if (now - lastUpdateTime.current >= delay) {
-      setThrottledValue(value)
-      lastUpdateTime.current = now
-    } else {
-      const timeoutId = setTimeout(
-        () => {
-          setThrottledValue(value)
-          lastUpdateTime.current = Date.now()
-        },
-        delay - (now - lastUpdateTime.current),
-      )
-      return () => clearTimeout(timeoutId)
-    }
-  }, [value, delay])
-
-  return throttledValue
 }
 
 export const useStatisticsTab = (): ReturnProps => {
